GEODE-7219 BufferUnderflowException in PutReplyMessage deserialization
VersionTag serialization was being affected by concurrent modification
of its memberId/previousMemberId fields, causing the HAS_PREVIOUS_MEMBER_ID
flag bit to be set and the DUPLICATE_MEMBER_IDS flag to _not_ be set.
It then went on to perform the same checks later in toData() and make
different decisions, winding up by not serializing the previousMemberId
field because it was then == to the memberId field.

1) decide on what the flags are going to be in toData and do not perform
the same calculations again.

2) use equals() when seeing if memberId and previousMemberId are equal.