GEODE-7237: Fix race condition in GfshRule
Creation of geode-log4j widened a race condition in GfshRule in which
the process being forked would start printing output before the
ProcessLogger had been created and started in order to intercept that
output.

The widening of that race condition resulted in CI failures of
ConnectCommandAcceptanceTest.

ProcessLogger does not appear to have any real need for manipulating
log4j-core which is what widens the race condition far enough to fail
in CI.

Removing the manipulation of log4j-core from ProcessLogger reduces
the race condition for consistent passing. The window itself is still
there but without resorting to something like MainLauncher [1], there's
no way to completing close the window.

[1] geode-junit: org.apache.geode.test.process.MainLauncher

If ConnectCommandAcceptanceTest or other tests using GfshRule fail
after this commit, then we'll need to use something like MainLauncher.