GEODE-5644: Add new DistributedReference JUnit Rule
JUnit Rule that provides a static reference in every distributed test
VMs including the main JUnit controller VM. If the referenced value is
is an AutoCloseable or Closeable then it will be auto-closed in every
VM during tear down.

If the referenced value is not an AutoCloseable or Closeable, the
DistributedReference will use reflection to invoke any method named
close or disconnect regardless of what interfaces are implemented.

If the referenced value is null in any VM then it will be ignored in
that VM during tear down.

To disable auto-close, simply specify autoClose(false).

See the javadocs on DistributedReference for more info.