GEODE-7283: Use no-op authorizer on client side
CQ tests were failing because the `getQueryService()` method on client
side (invoked when creating the new `ExecutionContext`) causes an
implicit `DEFAULT` Pool initialization, which can not be avoided when
using custom pools as the tests do.
The `ExecutionContext` now uses a no-op authorizer when created from a
client cache: authorization and authentication logic happens only on
server side anyways, and the authorizer will be obtained from the
cluster configuration service directly instead of the query service
when OQL security is fully implemented (GEODE-6983).