GEODE-5194: allow gfsh to connect if serialization version matches
This builds on the previous work that relaxed the version check to
ignore patch version changes.  Now even differing minor versions may be
compatible (i.e. no changes to gfsh commands or options).  Serialization
version is not a perfect proxy for when gfsh commands have changed, but
this is at least an improvement over the previous check...

Co-authored-by: Jinmei Liao <jiliao@pivotal.io>