GEODE-7400: Prevent RejectedExecutionException
Commit f0c96db73263bb1b3cb04558f2a720d70f43421f changed the
FederatingManager class so that it reuses the same ExecutorService
between restarts. After that change, if we start the manager after
previously starting and stopping it, we get RejectedExecutionException
because it tries to invoke a task on the same ExecutorService which has
been shut down.

This commit changes the FederatingManager so that it invokes a supplier
to get a new ExecutorService each time it is started to prevent the
RejectedExecutionException.

Co-authored-by: Aaron Lindsey <alindsey@pivotal.io>
Co-authored-by: Kirk Lund <klund@apache.org>