Adding a start method to the Membership API
The MembershipManagerAdapter class had a lifecycle problem where
membershipManager was null at the time that setDirectChannel was called,
because it happened during the construction of MembershipManagerAdapter.

This caused NPEs in integration tests.

To fix this, breaking out a start method from the create method of Membership.
Now the MembershipManagerAdapter has a separate constructor followed by a start
call.