Making QuorumChecker part of the membership API
The Membership interface was returning a QuorumChecker, but that was not part
of the API package. Moving that to API package. Extracting an interface for
MembershipInformation, which is opaque to consumers of the membership system -
it should be returned from a QuorumChecker and passed back into the new
membership system.