GEODE-7485: REST URLs accepts dots in trailing parameters
By default, when Spring parses a request mapping parameter at the end of
a URL, it treats the last dot in the parmater value to be the start of a
file extension, and discards the 'extension" before passing the
parameter to the handler method.

Several of our configurable items allow dots in their IDs. And several
of our REST request mappings place those IDs at the end of the URL. In
those cases, Spring's default behavior truncates the ID at the last dot,
and passes the truncated ID to the request handler. As a result, the
request fails.

This commit configure all current management REST API request mappings
to allow dots in their final parameters.

Co-authored-by: Dale Emery <demery@pivotal.io>
Co-authored-by: Jinmei Liao <jiliao@pivotal.io>
Co-authored-by: Dale Emery <demery@pivotal.io>