GEODE-7519 GMSHealthMonitorJUnitTest fails in testCheckIfAvailableNoHeartBeatDontRemoveMember
Removing time-sensitive test.  There are other tests that already
exercise the checkIfAvailable health-monitor method.  Two of those tests make
sure that the initiateRemoval parameter (the last parameter) to
checkIfAvailable functions properly and that the value returned by the
method is correct.  Other tests already ensure that a member that
checks out okay is not removed from the cluster and that the value
returned by checkIfAvailable is correct.