GEODE-7551: Remove membership API dependency on ClusterDistributionManager
ClusterDistributionManager was only being used to get canonical IDs and
to check for shutdown conditions.  I've modified the CDM to set shutdown
status in membership in its shutdown() method and have modified
membership to just use the current MembershipView (which contains any
surprise members) instead of asking CDM for a canonical ID, which just
grabs said MembershipView and does the same thing.