GEODE-7552: Break dependency on Services
Rewrote locator installation to use Membership rather than the "hidden"
Services interface.

I rewrote some tests to stop using the internal interfaces of membership
as well, and along the way saw that JGroupsMessenger was maintaining a
collection of old identifiers that was only being consumed by a test & I
removed it.