GEODE-7536: Add serialization analysis to geode-gfsh (#4446)
* GEODE-7536: Add serialization analysis to geode-gfsh

The new geode-gfsh module left all of its serializables in the
sanctioned list from geode-core.  This commit adds serialization
analysis to the new module and removes geode-gfsh serializables from
sanctioned-geode-core-serializables.txt.

I had to modify the base test class to read the sanctioned list using a
resource stream instead of a file.  Geode-gfsh, for some reason, has the
module's JAR file on the path and the test is finding the resource there
instead of as a regular file.

* resolving Dan's review

* removing unused instance variable