GEODE-7256: Move CacheFactoryStaticsTest to IntegrationTest
RemoteGfManagerAgentTest and CacheFactoryStaticsTest failed in my PR
precheckin, so I studied the tests and related changes that were
recently merged in.

Move CacheFactoryStaticsTest to IntegrationTest. Rename to
CacheFactoryStaticsIntegrationTest. This test is creating a full Cache
which makes it an IntegrationTest.

Remove unnecessary spy from RemoteGfManagerAgentTest.

Even though RemoteGfManagerAgentTest was setting the system field in
RemoteGfManagerAgent, the test intermittently nulls out the system and
then creates a real connection to the DistributedSystem which is not
desirable in a unit test, so I replaced static dependencies with
factories that can be stubbed to inject the desired mocks.

Cleanup RemoteGfManagerAgent, RemoteGfManagerAgentTest, and
CacheFactoryStaticsIntegrationTest. Fix up unchecked types and make all
fields private again. Add VisibleForTesting annotation to applicable
methods. Use constructor injection to avoid breaking encapsulation.