GEODE-7554: Add retry mechanism for failed tests (#4461)
* GEODE-7554: Add retry mechanism for failed tests

There are still some flaky benchmarks. Retry tests that have failed, up
to 5 times, to determine if the failure is legitimate or just a flaky
test. If the test fails 5 times in a row, we know that it is a
legitimate failure.

With each run of run_against_baseline (in the apache/geode-benchmarks
repo), the class names of all failed tests will be written to a file.
The run_benchmarks script called by CI will read the failed tests from
the file and run only those tests. Once the return code is 0, or once
we've tried 5 times, exit.

* build failedFile path dynamically

Signed-off-by: Helena A. Bales <hbales@pivotal.io>

* add print of iteration number for CI debugging of failures

Signed-off-by: Robert Houghton <rhoughton@pivotal.io>

* triggering CI again. Previous status never updated

Signed-off-by: Helena A. Bales <hbales@pivotal.io>

* refactor if statement for readability

* restarting CI

* remove hardcoded cluster tag

(cherry picked from commit 278c2470a5cac2c332d13914b935f0618b820a91)