GEODE-7627: Moving defaults to MembershipConfig
Moving the defaults of membership related settings to the
MembershipConfig class, and referencing them from DistributionConfig.
This makes it easier to create a Membership with the correct default
values.

Adding default implementations of many of the Membership colaborators,
such as AuthenticatorNoOp and MembershipStatisticsNoOp.

Adding a test of creating a Membership using only membership classes.
This test still has several TODOs related to core classes which were
necessary or helpful to make this work.