GEODE-7592: Simplify startManager() precondition checks
LGTM complained about a possible NPE in startManager(). There was no
possibility of an NPE, but precondition-checking code was overly
complex, and difficult for LGTM and humans to analyze.

It turns out that much of the complexity was due to the code checking
for impossible conditions.

This PR greatly simplifies the precondition checks, and adds a suite of
tests for the startManager() method.

Adding the tests required injecting several dependencies.

Co-authored-by: Dale Emery <demery@pivotal.io>
Co-authored-by: Joris Melchior <joris.melchior@gmail.com>