GEODE-7545: Remove dependency on AltertingAction from membership
The only place where JGroupsMessenger called AlertingAction was when
sending a message. The only message that could be sent while that thread
local is true is AlertListenerMessage.

Setting isHighPriority to true on AlertListenerMessage instead.
JGroupsMessenger is the only place that flag is checked, so this should
have the same effect as testing the thread local.