Removing hostname resolution from TcpSocketCreatorAdapter
We didn't need these methods on TcpSocketCreatorAdapter. resolveDns is based on
whether network partition detection is enabled, which we can access directly in
membership. getHostName uses a static hostname cache that we don't need to use
for membership. And getLocalHost has moved to a utility in geode- that we don't
need to use for membership. And getLocalHost has moved to a utility in
geode-common.