Fixing integration test failures
I moved the logic where the static Locator.hasLocator method is called. Some of
the tests were relying on this logic being in membership to realize that the
member was itself a locator. I changed those tests to specify that that are
using the LOCATOR_DM_TYPE.

Removing a call to restarting with a bunch of null fields. That call used to be
there and have no effect, but with the new RestartHandlers it was actually
passing the nulls down.