GEODE-7649 upgrade tests fail when v1.11 is added to geode-old-versions
Geode v1.11 inadvertently changed the form of its saved membership view
by using a FileOutputStream to write it to disk instead of using
an ObjectOutputStream.  The latter writes additional information during
serialization.

Note: there is already a test for this change that will start running
when v1.12 is created and v1.11 is added as a geode-old-version in
settings.gradle.