Check the resolve_dns flag before generating HA region queue name
Before the previous changes to InternalDistributedMember, we were flipping the
hostname field to be an IP address when we deserialized the IDM, based on the
resolve_dns flag.

It looks like this was required to keep the HA region queue name generated from
addFixedToString to match between old versions and the current version - the
region name should have an IP address if network partition detection is
enabled.