GEODE-7703: Catch IndexWriter Exceptions
The IndexWriter initialization might fail when other threads are
updating the fileAndChunkRegion, which can be triggered by other normal
operations (query, event listener, close, reindex, etc.). This doesn't
happen often and, instead of propagating the exception to the caller
and failing, Geode now catches it and returns null to let the callers
retry.

- Added unit and distrbuted tests.
- Return null instead of re-throwing the IOException while building
  the Lucene IndexWriter.

Co-authored-by: Xiaojian Zhou <gzhou@pivotal.io>