GEODE-7698: Extracting a superclass of InternalDistributedMember
GEODE-7698: Extracting a superclass of InternalDistributedMember

We need a concrete MemberIdentifier implementation in geode-membership
to test membership.

Extracting a superclass of InternalDistributedMember,
MemberIdentifierImpl, which contains pretty much all of the
InternalDistributedMember logic. InternalDistributedMember now just
extends that class and implements DistributedMember.

Removing unecessary use of SocketCreator during IDM deserialization. We were 
using flags and methods on SocketCreator to either change or cache the hostName
that we read off the wire from the remote member. We don't want to do either of
those things during deserialization. We should just keep the hostname we read
as is. This logic was used to change the toString value of ids that come from
clients to use the ip address instead of the hostname. Instead of changing the
deserialization logic, we change the addFixedToString logic.

Check for IP address rather than SocketCreator.resolve_dns in addFixedToString. 
InternalDistributedMember has a dependency on this mutable static flag in it's
to string, which is not ideal. It looks like the intention was to only shorten
the hostname if it is dns name, not an ip address.