Workaround for isCurrentServerMostLoaded
"LoadHolder isCurrentLH" should contain the current LoadHolder object
of the server we are connected to. But as we dont have the complete
key of the connectionMap (we dont have memberId) we have to do this
workaround to find it.
In case of several servers sharing host and port, it is not guaranteed
that we are returning the correct LoadHolder object.