GEODE-7656: OOME in CoreOnlyUsesMembershipAPIArchUnitTest > cacheClassesDoNotUseMembershipInternals
The ArchUnit class file importer's javadocs don't mention it but you
can't examine just a single package using a ClassFileImporter.  It
always walks the subpackages of any package you give it and includes all
of the classes found there.

I've split off the "org.apache.geode" check into its own method and have
added a custom ImportOption that uses a Pattern to reject anything in a
sub-package of org.apache.geode or in the membership package (which is
needed for the API check).  That reduced the number of classes being
examined from over 7700 to less than 300.