GEODE-7760: fixing LocatorUDPSecurityDUnitTest.testCrashLocatorMultipleTimes
This new test exposed a flaw in auto-reconnect where the quorum-checker
buffers messages from other members but the new Membership messenger is
unable to decrypt them when it starts up due to not having the old
encryption keys.

The encrypt/decrypt objects from the old Membership are now kept and
provided to the new Messenger during auto-reconnect.