GEODE-7866: Clean up geode-tcp-server module and add missing javadocs
Added package-info javadoc

Renamed ServerSocketCreatorImpl to ClusterSocketCreatorImpl and renamed
variables associated with these objects

Removed TcpServerGossipVersionDUnitTest, which was only needed in old
pre-Geode code when transitioning from GemFire 6x to GemFire 7x.

Added javadocs to classes and public methods.

Changed TcpServerProductVersionDUnitTest to use a TcpServer instead of a
Locator.  Modified the test to wait for the TcpServer to shut down after
it receives a ShutdownMessage.