GEODE-7830: fix race in rebalance start (#4769)
* The result from doing a start will now be consistent.
Previously the OperationState used to compute the result
could be changing while making the result causing the result
to be inconsistent. Now a copy of the OperationState is made
in a thread safe manner and then that immutable copy is used
to compute the result.

* A snapshot of the operationState is now taken BEFORE
the async thread is started so that the result of "start"
will only have info about the start.

* Now if a rebalance is done of the entire cluster it
will no longer be an error if it does not find any
regions to rebalance. Instead it will report success
with the status message: "Distributed system has no regions that can be rebalanced".
Also if it a member departs during the rebalance such that a region
it was hosting can no longer be rebalanced, then that region is now
ignored by rebalance instead of an error being reported.