GEODE-7874 write a serialization backward-compatibility test for geode-membership
Split the TestBase class into two classes, one which analyzes
DataSerializable-like classes and a subclass that analyzes
java-serializable classes.

Added an AnalyzeMembershipSerializablesJUnitTest that implements the
first TestBase and left other AnalyzeSerializables tests as subclasses
of the second (AnalyzeSerializablesJUnitTestBase).

I also finally moved the AnalyzeManagementSerializablesJUnitTest into
the geode-management module, where it belongs.  It wasn't properly
testing files in that module before and I had to update its sanctioned
list with a couple of new Exceptions that Owen created a while ago.

k