GEODE-7877: deal with the static Version map in TcpClient
The server-version map that tracks Locator versions is now an instance
variable instead of a static.

I looked into removing the map altogether but it's really needed for
rolling-upgrade purposes.  When one locator is rolled to a new version
of Geode but another locator is still running the old version it's
essential that the new locator know that it's communicating with an old
locator.