GEODE-7900: Only set the SNI hostname if it is not present
It turns out the JDK sometimes automatically sets the SNI hostname and
sometimes does not.  With openjdk 8, it only sets SNI hostname if there is a
dot in the hostname. See sun.security.ssl.Utilities#addToSNIServerNameList

For this reason, we need to only set the SNI hostname if the JDK did not.
Otherwise it will complain that there is a `Duplicated server name of type 0`